% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_rendas_brasileirao.R
\name{baixa_rendas_brasileirao}
\alias{baixa_rendas_brasileirao}
\title{baixa as rendas dos jogos do brasileirão}
\usage{
baixa_rendas_brasileirao(n_rodadas, ano)
}
\arguments{
\item{n_rodadas}{vetor com número de rodadas}

\item{ano}{ano do campeonato}
}
\value{
dataframe com os dados sobre o financiamento das partidas do brasileirão
}
\description{
baixa as rendas dos jogos do brasileirão
}
\examples{
baixa_rendas_brasileirao(1:15, 2022)
}
