% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_rendas_brasileirao_b.R
\name{baixa_rendas_brasileirao_b}
\alias{baixa_rendas_brasileirao_b}
\title{baixa_rendas_brasileirao_b}
\usage{
baixa_rendas_brasileirao_b(n_rodadas, ano)
}
\arguments{
\item{n_rodadas}{vetor com número de rodadas}

\item{ano}{ano do campeonato}
}
\value{
dataframe com os dados sobre o financiamento das partidas da série B do brasileirão
}
\description{
baixa_rendas_brasileirao_b
}
\examples{
baixa_rendas_brasileirao_b(1:15, 2015)
}
