% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_rendas_brasileirao_c.R
\name{baixa_rendas_brasileirao_c}
\alias{baixa_rendas_brasileirao_c}
\title{baixa_rendas_brasileirao_c}
\usage{
baixa_rendas_brasileirao_c(n_rodadas)
}
\arguments{
\item{n_rodadas}{vetor com número de rodadas}
}
\value{
dataframe com os dados sobre o financiamento das partidas da série C do brasileirão
}
\description{
baixa_rendas_brasileirao_c
}
\examples{
baixa_rendas_brasileirao_c(1:15)
}
