% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_rendas_brasileirao_d.R
\name{baixa_rendas_brasileirao_d}
\alias{baixa_rendas_brasileirao_d}
\title{baixa_rendas_brasileirao_d}
\usage{
baixa_rendas_brasileirao_d(n_rodadas)
}
\arguments{
\item{n_rodadas}{vetor com número de rodadas}
}
\value{
dataframe com os dados sobre o financiamento das partidas da série C do brasileirão
}
\description{
baixa_rendas_brasileirao_d
}
\examples{
baixa_rendas_brasileirao_d(1:15)
}
