% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_tabela_serie_d.R
\name{cria_tabela_serie_d}
\alias{cria_tabela_serie_d}
\title{Cria um datavis em formato tabular das medidas de variacao para os Clubes da Série B}
\usage{
cria_tabela_serie_d(
  df,
  variavel,
  clubem,
  placarm_tn,
  placarv_tn,
  titulo = NULL
)
}
\arguments{
\item{df}{dataset criado a a partir da funcao baixa_rendas_brasileirao}

\item{variavel}{variavel  numerica que provem os dados da tabela}

\item{clubem}{clube mandante}

\item{placarm_tn}{placar do clube mandante}

\item{placarv_tn}{placar do clube visitante}

\item{titulo}{string de texto que compoe o titulo da tabela}
}
\value{
a tibble
}
\description{
Cria um datavis em formato tabular das medidas de variacao para os Clubes da Série B
}
\examples{

}
