% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_id_time.R
\name{extrai_id_time}
\alias{extrai_id_time}
\title{extrai id dos times da serie A}
\usage{
extrai_id_time(ano)
}
\arguments{
\item{ano}{do campeonato}
}
\value{
um dataframe com os ids dos times da serie A
}
\description{
extrai id dos times da serie A
}
