% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_estatisticas.R
\name{pega_estatisticas}
\alias{pega_estatisticas}
\title{pega as estatitisticas dos times da serie a}
\usage{
pega_estatisticas(ano_campeonato)
}
\arguments{
\item{ano_campeonato}{ano do campeonato para retornar as estatisticas}
}
\value{
um dataframe as estatisticas dos times da serie
}
\description{
pega as estatitisticas dos times da serie a
}
